Move colours and the colour functions to their own file
