+++ src/renderer/components/theme-settings/theme-settings.js
@@ -6,7 +6,7 @@ import FtToggleSwitch from '../ft-toggle-switch/ft-toggle-switch.vue'
 import FtSlider from '../ft-slider/ft-slider.vue'
 import FtFlexBox from '../ft-flex-box/ft-flex-box.vue'
 import FtPrompt from '../ft-prompt/ft-prompt.vue'
-import { colors } from '../../helpers/utils'
+import { colors } from '../../helpers/colors'
 
 export default Vue.extend({
   name: 'ThemeSettings',
