+++ src/renderer/helpers/utils.js
@@ -5,71 +5,6 @@ import FtToastEvents from '../components/ft-toast/ft-toast-events'
 import i18n from '../i18n/index'
 import router from '../router/index'
 
-export const colors = [
-  { name: 'Red', value: '#d50000' },
-  { name: 'Pink', value: '#C51162' },
-  { name: 'Purple', value: '#AA00FF' },
-  { name: 'DeepPurple', value: '#6200EA' },
-  { name: 'Indigo', value: '#304FFE' },
-  { name: 'Blue', value: '#2962FF' },
-  { name: 'LightBlue', value: '#0091EA' },
-  { name: 'Cyan', value: '#00B8D4' },
-  { name: 'Teal', value: '#00BFA5' },
-  { name: 'Green', value: '#00C853' },
-  { name: 'LightGreen', value: '#64DD17' },
-  { name: 'Lime', value: '#AEEA00' },
-  { name: 'Yellow', value: '#FFD600' },
-  { name: 'Amber', value: '#FFAB00' },
-  { name: 'Orange', value: '#FF6D00' },
-  { name: 'DeepOrange', value: '#DD2C00' },
-  { name: 'DraculaCyan', value: '#8BE9FD' },
-  { name: 'DraculaGreen', value: '#50FA7B' },
-  { name: 'DraculaOrange', value: '#FFB86C' },
-  { name: 'DraculaPink', value: '#FF79C6' },
-  { name: 'DraculaPurple', value: '#BD93F9' },
-  { name: 'DraculaRed', value: '#FF5555' },
-  { name: 'DraculaYellow', value: '#F1FA8C' },
-  { name: 'CatppuccinMochaRosewater', value: '#F5E0DC' },
-  { name: 'CatppuccinMochaFlamingo', value: '#F2CDCD' },
-  { name: 'CatppuccinMochaPink', value: '#F5C2E7' },
-  { name: 'CatppuccinMochaMauve', value: '#CBA6F7' },
-  { name: 'CatppuccinMochaRed', value: '#F38BA8' },
-  { name: 'CatppuccinMochaMaroon', value: '#EBA0AC' },
-  { name: 'CatppuccinMochaPeach', value: '#FAB387' },
-  { name: 'CatppuccinMochaYellow', value: '#F9E2AF' },
-  { name: 'CatppuccinMochaGreen', value: '#A6E3A1' },
-  { name: 'CatppuccinMochaTeal', value: '#94E2D5' },
-  { name: 'CatppuccinMochaSky', value: '#89DCEB' },
-  { name: 'CatppuccinMochaSapphire', value: '#74C7EC' },
-  { name: 'CatppuccinMochaBlue', value: '#89B4FA' },
-  { name: 'CatppuccinMochaLavender', value: '#B4BEFE' }
-]
-
-export function getRandomColorClass() {
-  const randomInt = Math.floor(Math.random() * colors.length)
-  return 'main' + colors[randomInt].name
-}
-
-export function getRandomColor() {
-  const randomInt = Math.floor(Math.random() * colors.length)
-  return colors[randomInt].value
-}
-
-export function calculateColorLuminance(colorValue) {
-  const cutHex = colorValue.substring(1, 7)
-  const colorValueR = parseInt(cutHex.substring(0, 2), 16)
-  const colorValueG = parseInt(cutHex.substring(2, 4), 16)
-  const colorValueB = parseInt(cutHex.substring(4, 6), 16)
-
-  const luminance = (0.299 * colorValueR + 0.587 * colorValueG + 0.114 * colorValueB) / 255
-
-  if (luminance > 0.5) {
-    return '#000000'
-  } else {
-    return '#FFFFFF'
-  }
-}
-
 export function calculatePublishedDate(publishedText) {
   const date = new Date()
   if (publishedText === 'Live') {
