+++ src/renderer/views/ProfileEdit/ProfileEdit.js
@@ -5,7 +5,8 @@ import FtProfileEdit from '../../components/ft-profile-edit/ft-profile-edit.vue'
 import FtProfileChannelList from '../../components/ft-profile-channel-list/ft-profile-channel-list.vue'
 import FtProfileFilterChannelsList from '../../components/ft-profile-filter-channels-list/ft-profile-filter-channels-list.vue'
 import { MAIN_PROFILE_ID } from '../../../constants'
-import { calculateColorLuminance, getRandomColor, showToast } from '../../helpers/utils'
+import { calculateColorLuminance, getRandomColor } from '../../helpers/colors'
+import { showToast } from '../../helpers/utils'
 
 export default Vue.extend({
   name: 'ProfileEdit',
